# api.yaml
swagger: '2.0'
info:
  title: API микросервиса общих справочников
  version: 1.0.0
paths:
...
  /dictionaries/currencies/{id}:
    get:
      tags:
        - Currencies
      description: Возвращает информацию о валюте
      operationId: getCurrency
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор валюты
      responses:
        200:
          description: ОК
          schema:
            description: Информация о валюте
            $ref: '#/definitions/currency'
        404:
          description: Not Found
    put:
      tags:
        - Currencies
      description: Редактирование валюты
      operationId: editCurrency
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор валюты
        - in: body
          name: currency
          description: Данные о валюте
          required: true
          schema:
            $ref: '#/definitions/currency'
      responses:
        200:
          description: ОК
          schema:
            description: Сохраненная валюта
            $ref: '#/definitions/currency'
        404:
          description: Not Found
    delete:
      tags:
        - Currencies
      description: Удаление валюты
      operationId: deleteCurrency
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор валюты
      responses:
        204:
          description: No Content
        404:
          description: Not Found

  #***currency-rate***
  /dictionaries/rates/setting:
    get:
      tags:
      - CurrencyRate
      description: Получить насройки способа расчета курса валют
      operationId: getCurrencyRateSetting
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            description: Ответ на запрос получения настроек способа расчета курса валют
            $ref: '#/definitions/currencyRateSettings'
        403:
          description: Forbiden
        500:
          description: Internal Server Error
    post:
      tags:
        - CurrencyRate
      description: Обновить настройки способа расчета курса валют
      operationId: updateCurrencyRateSetting
      consumes:
        - application/json
      parameters:
        - in: body
          name: updateCurrencyRateSetting
          description: Данные с настройками способа расчета курса
          required: true
          schema:
            $ref: '#/definitions/currencyRateSettings'
      responses:
        200:
          description: OK
        403:
          description: Forbiden
        500:
          description: Internal Server Error
  /dictionaries/rates:
    get:
      tags:
        - CurrencyRate
      description: Получить постраничный список курсов валют для отображения в списке
      operationId: getCurrencyRateList
      x-spring-pageable: true
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            description: Постраничный вывод курсов валют
            $ref: '#/definitions/currencyRateDataForList'
        403:
          description: Forbidden
        500:
          description: Internal server error
    post:
      tags:
        - CurrencyRate
      description: Создать запись в справочнике курсов валют
      operationId: createCurrencyRate
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: createRateRequest
          description: Запрос создания записи справочника
          required: true
          schema:
            $ref: '#/definitions/currencyRateData'
      responses:
        200:
          description: OK
        403:
          description: Forbiden
        500:
          description: Internal Server Error
  /dictionaries/rate/{id}:
    get:
      tags:
        - CurrencyRate
      description: Получить данные курса по идентификатору записи в справочнике
      operationId: getCurrencyById
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          description: Идентификатор записи справочника
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/currencyRateData'
            description: Данные курса
        403:
            description: Forbidden
        404:
          description: Not Found
        500:
          description: Internal server error
    put:
      tags:
        - CurrencyRate
      operationId: updateCurrencyRate
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          description: Идентификатор записи справочника
          required: true
        - in: body
          name: updateRateRequest
          description: Запрос обновления записи справочника
          required: true
          schema:
            $ref: '#/definitions/currencyRateData'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/currencyRateData'
            description: Данные курса
        403:
          description: Forbidden
        404:
          description: Not Found
        500:
          description: Internal server error
  /dictionaries/currency/rate/cb/:
    get:
      tags:
        - CurrencyRate
      description: Получить список курсов валют по ЦБ РФ
      operationId: getCbCurrencyRateList
      produces:
        - application/json
      parameters:
        - in: query
          name: literalIso
          type: string
          description: ISO-код валюты
          required: true
        - in: query
          name: startDate
          description: Дата, с которой необходимо получить курсы. Формат ДД.ММ.ГГГГ
          required: true
          type: string
        - in: query
          name: endDate
          description: Дата, по которую необходимо получить курсы. Формат ДД.ММ.ГГГГ
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/currencyCbRateRs'
            description: Курс валюты
        404:
          description: Not Found
  /dictionaries/currency/rate/active:
    get:
      tags:
        - CurrencyRate
      description: Получить активный курс валюты по идентификатору валюты
      operationId: getActiveCurrencyRate
      produces:
        - application/json
      parameters:
        - in: query
          name: currencyId
          type: integer
          format: int64
          description: Идентификатор валюты
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/shortCurrencyRateData'
            description: Курс валюты
        404:
          description: Not Found
  #***external-systems***
...
definitions:
  currencyCbRateRs:
    type: object
    description: Ответ на запрос получения курсов валют ЦБ РФ
    properties:
      rates:
        type: array
        items:
          $ref: '#/definitions/currencyCbRateData'
  currencyCbRateData:
    type: object
    description: Курс валюты ЦБ РФ
    required:
      - currencyIso
      - date
      - rate
    properties:
      currencyIso:
        type: string
        description: ISO-код валюты
      date:
        type: string
        description: Дата действия курса. Формат ДД.ММ.ГГГГ
      rate:
        type: number
        description: Курс ЦБ РФ
  currencyRateSettings:
    type: object
    description: Настройки расчета курса валюты
    properties:
      calculationType:
        $ref: '#/definitions/calculationType'
        description: Способ расчета курса валюты Организации
      additionalPercent:
        type: number
        description: Добавочный процент для расчета курса валюты организации
  currencyRateDataForList:
    allOf:
      - $ref: '#/definitions/currencyRateData'
      - type: object
        properties:
          currencyIso:
            type: string
            description: ISO-код валюты
  currencyRateData:
    type: object
    description: Курс валюты
    required:
      - currencyId
      - rate
    properties:
      id:
        type: integer
        format: int64
        description: Идентификатор записи справочника
      currencyId:
        type: integer
        format: int64
        description: Идентификатор валюты
      rate:
        type: number
        description: Курс ЦБ РФ
      innerRate:
        type: number
        description: Курс Организации
      startDate:
        type: string
        format: date-time
        description: Дата начала действия курса
      endDate:
        type: string
        format: date-time
        description: Дата окончания действия курса
      active:
        type: boolean
        description: Признак действующего курса валюты
      edited:
        type: boolean
        description: Признак ручного редактирования
  shortCurrencyRateData:
    type: object
    description: Курс валюты
    required:
      - id
      - currencyId
      - rate
    properties:
      id:
        type: integer
        format: int64
        description: Идентификатор записи справочника
      currencyId:
        type: integer
        format: int64
        description: Идентификатор валюты
      rate:
        type: number
        description: Курс ЦБ РФ
      innerRate:
        type: number
        description: Курс Организации
...
  calculationType:
    type: string
    description: Способ расчета курса валюты Организации
    enum:
      - CB_RF_RATE
      - SET_BY_ADMINISTRATOR
      - CB_RF_WITH_PERCENT
...
  currency:
    type: object
    description: Валюта
    properties:
      id:
        type: integer
        format: int64
        description: Id записи
      literalISO:
        type: string
        description: ISO код
      digitalISO:
        type: integer
        format: int64
        description: Цифровой код
      currencyName:
        type: string
        description: Наименование
      fullCurrencyName:
        type: string
        description: Полное наименование
      countryHolder:
        type: string
        description: Страна-эмиссионер
...